+++ vortex-scalar/src/bool.rs
@@ -1,38 +1,29 @@
-use std::fmt::{Display, Formatter};
-
-use vortex_dtype::{DType, Nullability};
+use vortex_dtype::DType;
 use vortex_error::{vortex_bail, vortex_err, VortexError, VortexResult};
 
-use crate::value::ScalarValue;
 use crate::Scalar;
 
-pub type BoolScalar = ScalarValue<bool>;
-
-impl BoolScalar {
+pub struct BoolScalar<'a>(&'a Scalar);
+impl<'a> BoolScalar<'a> {
     #[inline]
-    pub fn dtype(&self) -> &DType {
-        match self.nullability() {
-            Nullability::NonNullable => &DType::Bool(Nullability::NonNullable),
-            Nullability::Nullable => &DType::Bool(Nullability::Nullable),
-        }
+    pub fn dtype(&self) -> &'a DType {
+        self.0.dtype()
     }
 
-    pub fn cast(&self, dtype: &DType) -> VortexResult<Scalar> {
-        match dtype {
-            DType::Bool(_) => Ok(self.clone().into()),
-            _ => Err(vortex_err!(MismatchedTypes: "bool", dtype)),
-        }
-    }
-
-    pub fn nbytes(&self) -> usize {
-        1
+    pub fn value(&self) -> Option<bool> {
+        self.0.value.as_bool()
     }
 }
 
-impl From<bool> for Scalar {
-    #[inline]
-    fn from(value: bool) -> Self {
-        BoolScalar::some(value).into()
+impl<'a> TryFrom<&'a Scalar> for BoolScalar<'a> {
+    type Error = VortexError;
+
+    fn try_from(value: &'a Scalar) -> Result<Self, Self::Error> {
+        if matches!(value.dtype(), DType::Bool(_)) {
+            Ok(Self(value))
+        } else {
+            vortex_bail!("Expected bool scalar, found {}", value.dtype())
+        }
     }
 }
 
@@ -40,33 +31,9 @@ impl TryFrom<&Scalar> for bool {
     type Error = VortexError;
 
     fn try_from(value: &Scalar) -> VortexResult<Self> {
-        let Scalar::Bool(b) = value else {
-            vortex_bail!(MismatchedTypes: "bool", value.dtype());
-        };
-        b.value()
-            .cloned()
+        BoolScalar::try_from(value)?
+            .value()
             .ok_or_else(|| vortex_err!("Can't extract present value from null scalar"))
-    }
-}
-
-impl TryFrom<Scalar> for bool {
-    type Error = VortexError;
-
-    fn try_from(value: Scalar) -> VortexResult<Self> {
-        let Scalar::Bool(b) = value else {
-            vortex_bail!(MismatchedTypes: "bool", value.dtype());
-        };
-        b.into_value()
-            .ok_or_else(|| vortex_err!("Can't extract present value from null scalar"))
-    }
-}
-
-impl Display for BoolScalar {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self.value() {
-            None => write!(f, "null"),
-            Some(b) => Display::fmt(&b, f),
-        }
     }
 }
 
@@ -77,6 +44,6 @@ mod test {
     #[test]
     fn into_from() {
         let scalar: Scalar = false.into();
-        assert!(!bool::try_from(scalar).unwrap());
+        assert!(!bool::try_from(&scalar).unwrap());
     }
 }
