--- vortex-scalar/src/scalar2/mod.rs
@@ -1,46 +1,0 @@
-#![allow(dead_code)]
-
-use value::*;
-use vortex_dtype::DType;
-use vortex_dtype::Nullability::NonNullable;
-
-mod binary;
-mod bool;
-mod display;
-mod primitive;
-mod struct_;
-mod utf8;
-mod value;
-
-#[derive(Debug, Clone)]
-pub struct Scalar {
-    pub(crate) dtype: DType,
-    pub(crate) value: ScalarValue,
-}
-
-impl Scalar {
-    pub fn dtype(&self) -> &DType {
-        &self.dtype
-    }
-
-    pub fn is_null(&self) -> bool {
-        self.value.is_null()
-    }
-
-    pub fn null(&self, dtype: DType) -> Self {
-        assert!(dtype.is_nullable());
-        Self {
-            dtype,
-            value: ScalarValue::Data(ScalarData::None),
-        }
-    }
-}
-
-impl From<bool> for Scalar {
-    fn from(value: bool) -> Self {
-        Scalar {
-            dtype: DType::Bool(NonNullable),
-            value: ScalarValue::Data(ScalarData::Bool(value)),
-        }
-    }
-}
