--- vortex-scalar/src/scalar2/utf8.rs
@@ -1,47 +1,0 @@
-use vortex_buffer::BufferString;
-use vortex_dtype::DType;
-use vortex_error::{vortex_bail, vortex_err, VortexError, VortexResult};
-
-use crate::scalar2::Scalar;
-
-pub struct Utf8Scalar<'a>(&'a Scalar);
-impl<'a> Utf8Scalar<'a> {
-    #[inline]
-    pub fn dtype(&self) -> &'a DType {
-        self.0.dtype()
-    }
-
-    pub fn value(&self) -> Option<BufferString> {
-        self.0
-            .value
-            .as_bytes()
-            // Checked on construction that the buffer is valid UTF-8
-            .map(|buffer| unsafe { BufferString::new_unchecked(buffer) })
-    }
-}
-
-impl<'a> TryFrom<&'a Scalar> for Utf8Scalar<'a> {
-    type Error = VortexError;
-
-    fn try_from(value: &'a Scalar) -> Result<Self, Self::Error> {
-        if matches!(value.dtype(), DType::Utf8(_)) {
-            // Validate here that the buffer is indeed UTF-8
-            if let Some(buffer) = value.value.as_bytes() {
-                let _ = std::str::from_utf8(buffer.as_ref())?;
-            }
-            Ok(Self(value))
-        } else {
-            vortex_bail!("Expected utf8 scalar, found {}", value.dtype())
-        }
-    }
-}
-
-impl<'a> TryFrom<&'a Scalar> for BufferString {
-    type Error = VortexError;
-
-    fn try_from(value: &'a Scalar) -> VortexResult<Self> {
-        Utf8Scalar::try_from(value)?
-            .value()
-            .ok_or_else(|| vortex_err!("Can't extract present value from null scalar"))
-    }
-}
